Fixed Todo in AccountController
When we delete an Account we also delete the users belonging to the
account and the Brand (only if that account was the only one linking to
the Brand)

Fixed an issue with the ash tickets, when an account is deleted, the
ash ticket link gave an error (it couldn’t find the brand). When the
account in the ticket doesn’t exist anymore the ash controller will use
the default brand instead of the brand linked to the account.

#AIO-21