Change for classifications
All classifications are now being indexed by a KEY (string:100) instead of a integer. This creates a better configuration setup and allowes for class names to be multilingual. This change is not backwards compatible and needs a migrate:refresh (--seed). most parsers need version 1.2.x to be compatible with this change