Fixed changing ASH key
The ash key was calculated before saving using the ticket id, however
the id doesn’t exist when the ticket is created, so the second save
will change the ash tokens. If the contact of the ticket is notified,
he will have received an invalid ashtoken.

We now update the ash token after the ticket is saved, so it will
always have an id, and the ash ticket isn’t build anymore with an empty
id